requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
multidict==6.0.4
orjson==3.9.7
owjcommon==0.1.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
PyJWT==2.8.0
pyotp==2.9.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
smsir-python==1.0.7
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
twilio==8.9.1
typing-extensions==4.8.0
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
